rdma: Move eager detection to device creation
In prep for a future patch that makes more device things const, move
the eager support check from get_properties() to RDMA's device
creation, so that the rdma get_properties() call doesn't try to
modify the device structure.
