+++ src/nccl_ofi_rdma.cpp
@@ -14,6 +14,11 @@
 #include <unistd.h>
 #include <pthread.h>
 #include <stdlib.h>
+#ifdef HAVE_RDMA_FI_EXT_EFA_H
+#ifdef HAVE_RDMA_FI_EXT_EFA_H
+#include <rdma/fi_ext_efa.h>
+#endif
+#endif
 
 #include "nccl_ofi.h"
 #include "nccl_ofi_log.h"
@@ -418,7 +423,6 @@ int nccl_net_ofi_rdma_device_t::get_properties(nccl_ofi_properties_t *props)
 		static_assert(NCCL_OFI_MAX_RECVS <= (1 << NCCL_OFI_RDMA_RECV_IDX_BITS),
 					  "NCCL_OFI_MAX_RECVS must fit in RECV_IDX_BITS");
 		props->max_communicators = NCCL_OFI_RDMA_MAX_COMMS;
-		this->eager_support = props->eager_support;
 	} else {
 		return ret;
 	}
@@ -7075,6 +7079,22 @@ nccl_net_ofi_rdma_device_t::nccl_net_ofi_rdma_device_t(nccl_net_ofi_plugin_t *pl
 			      strerror(-ret));
 		throw std::runtime_error("RDMA device constructor: connection prep failed");
 	}
+
+	/* Verify support for eager messages */
+#if HAVE_DECL_FI_EFA_FEATURE_OPS
+	{ /* Scope block: C++ forbids goto past auto variable initialization */
+		struct fi_efa_feature_ops *feat_ops = NULL;
+		ret = fi_open_ops(&this->device_rails[0].fabric->fid, FI_EFA_FEATURE_OPS, 0,
+				  (void **)&feat_ops, NULL);
+		if (ret != 0) {
+			NCCL_OFI_WARN("fi_open_ops for EFA features failed. RC: %d, ERROR: %s",
+				      ret, fi_strerror(-ret));
+			ret = 0;
+		} else if (feat_ops->query("mixed_hmem_iov")) {
+			this->eager_support = true;
+		}
+	}
+#endif
 
 	/* NVTX domain */
 #if HAVE_NVTX_TRACING
